mem: Document OOM behavior for allocations
For all the memory allocator APIS, document
that they terminate the program on failure.

This was so far only mentioned in the long description,
and in the docs for g_try_malloc(). And with gi-docgen
style docs, the long description is going away.

(cherry-picked from commit 1df83acb876c8325dfadf96bafc5751e1d9e5447)