ghash: document g_hash_table_steal_extended() behavior for sets
GHashTable is optimized for using g_hash_table_add(), where the
key and the value are identical.

When using g_hash_table_steal_extended() on such a dictionary,
and requesting both a key and a value, then the stolen value will be
NULL.

That would be easy to fix, but it also would be an API break.
Instead, document this behavior.