appinfo: Add new entries at the bottom
The state purpose of !3021 is to add a new fallback to `kgx`. This should not entail changing the order of the existing entries, which changes the order in which terminals are searched for. Move the new entry to the bottom, where fallback will still find it without changing the fallback order.